refactor sympy_expression
