--- neuralpp/symbolic/__init__.py
@@ -1,1 +1,0 @@
-from .functions import (conditional, )
