+++ neuralpp/symbolic/sympy_expression.py
@@ -1,172 +1,51 @@
 from __future__ import annotations
 
-import typing
-from typing import List, Any, Type, Callable, Tuple, Optional, Dict
 from abc import ABC, abstractmethod
+import fractions
+import operator
+from typing import Any, Callable, Dict, List, Optional, Tuple, Type
 
 import sympy
 from sympy import abc
-import operator
-import builtins
-import fractions
 
-from functools import cached_property
 from neuralpp.symbolic.expression import (
+    AbelianOperation,
+    Constant,
+    Context,
     Expression,
     FunctionApplication,
-    Variable,
-    Constant,
-    return_type_after_application,
-    Context,
     QuantifierExpression,
-    AbelianOperation,
+    Variable,
 )
-from neuralpp.symbolic.util import (
+from neuralpp.symbolic.functions import conditional
+from neuralpp.symbolic.parameters import global_parameters
+from neuralpp.util.callable_util import (
     ExpressionType,
-     infer_python_callable_type,
+    infer_sympy_function_type,
+    infer_sympy_object_type,
+    python_callable_to_sympy_function,
+    sympy_function_to_python_callable,
     return_type_after_application,
 )
-from neuralpp.symbolic.error import FunctionNotTypedError, NotTypedError, UnknownError
+from neuralpp.util.symbolic_error_util import (
+    FunctionNotTypedError,
+    NotTypedError,
+    UnknownError,
+)
+from neuralpp.util.sympy_util import (
+    fold_sympy_piecewise,
+    is_sympy_uninterpreted_function,
+    is_sympy_integral,
+    is_sympy_value,
+    is_sympy_sum,
+    sympy_piecewise_from_if_then_else,
+    sympy_piecewise_to_if_then_else,
+)
 from neuralpp.util.util import update_consistent_dict
-from neuralpp.symbolic.parameters import global_parameters
-import neuralpp.symbolic.functions as functions
-from neuralpp.util.sympy_util import is_sympy_uninterpreted_function
 
 
 # In this file's doc, I try to avoid the term `sympy expression` because it could mean both sympy.Expr (or sympy.Basic)
 # and SymPyExpression. I usually use "sympy object" to refer to the former and "expression" to refer to the latter.
-
-
-def _infer_sympy_object_type(
-    sympy_object: sympy.Basic, type_dict: Dict[sympy.Basic, ExpressionType]
-) -> ExpressionType:
-    """
-    type_dict can be, for example, {a: int, b: int, c: float, f: int->int}.
-    """
-    match sympy_object:
-        case sympy.Integer():
-            return int
-        case sympy.Float():
-            return float
-        case sympy.Rational():
-            return fractions.Fraction
-        case sympy.logic.boolalg.BooleanAtom():
-            return bool
-        case sympy.Piecewise():  # piecewise is not mapped in python_callable_and_sympy_function_relation
-            return _infer_sympy_object_type(sympy_object.args[0][0], type_dict)
-        case _:
-            # We can look up type_dict for variable like `symbol("x")`.
-            # A variable can also be an uninterpreted function `sympy.core.function.UndefinedFunction('f')`
-            try:
-                return type_dict[sympy_object]
-            except KeyError:  # if it's not in type_dict, try figure out ourselves
-                # Here, sympy_object must be function applications like 'x+y'
-                if (
-                    len(sympy_object.args) == 0
-                ):  # len(sympy_object.args) could raise (e.g, len(sympy.Add.args))
-                    raise TypeError("expect function application")
-                _, return_type = typing.get_args(
-                    _infer_sympy_function_type(sympy_object, type_dict)
-                )
-                return return_type
-
-
-def _infer_sympy_function_type(
-    sympy_object: sympy.Basic, type_dict: Dict[sympy.Basic, ExpressionType]
-) -> Callable:
-    """
-    Assume sympy_object is a function application, return the function type of the function application.
-    Note this is different from infer_sympy_object_type, which returns the type of function application.
-    E.g.,
-    >>> from sympy.abc import a, b
-    >>> _infer_sympy_function_type(a+b, {a:int, b:int})
-    Callable[[int, int], int]
-    >>> _infer_sympy_object_type(a+b, {a:int, b:int})
-    int
-    """
-    return infer_python_callable_type(
-        _sympy_function_to_python_callable(sympy_object.func),
-        [_infer_sympy_object_type(arg, type_dict) for arg in sympy_object.args],
-    )
-
-
-sympy_Sub = sympy.Lambda((abc.x, abc.y), abc.x - abc.y)
-sympy_Neg = sympy.Lambda((abc.x,), -abc.x)  # "lambda x: (-1)*x"
-sympy_Cond = sympy.Lambda(
-    (abc.i, abc.t, abc.e), sympy.Piecewise((abc.t, abc.i), (abc.e, True))
-)
-sympy_Div = sympy.Lambda((abc.x, abc.y), abc.x / abc.y)
-# Refer to sympy_simplification_test:test_unevaluate() for this design that uses sympy.Lambda()
-python_callable_and_sympy_function_relation = [
-    # boolean operation
-    (operator.and_, sympy.And),
-    (operator.or_, sympy.Or),
-    (operator.invert, sympy.Not),
-    (operator.xor, sympy.Xor),
-    # comparison
-    (operator.le, sympy.Le),
-    (operator.lt, sympy.Lt),
-    (operator.ge, sympy.Ge),
-    (operator.gt, sympy.Gt),
-    (operator.eq, sympy.Eq),
-    (operator.ne, sympy.Ne),
-    # arithmetic
-    (operator.add, sympy.Add),
-    (operator.mul, sympy.Mul),
-    (operator.pow, sympy.Pow),
-    (operator.sub, sympy_Sub),
-    (operator.neg, sympy_Neg),
-    (operator.truediv, sympy_Div),
-    # min/max
-    (builtins.min, sympy.Min),
-    (builtins.max, sympy.Max),
-    # conditional
-    (functions.conditional, sympy_Cond),
-]
-sympy_function_python_callable_dict = {
-    sympy_function: python_callable
-    for python_callable, sympy_function in python_callable_and_sympy_function_relation
-}
-python_callable_sympy_function_dict = {
-    python_callable: sympy_function
-    for python_callable, sympy_function in python_callable_and_sympy_function_relation
-}
-
-
-def _sympy_function_to_python_callable(sympy_function: sympy.Basic) -> Callable:
-    try:
-        return sympy_function_python_callable_dict[sympy_function]
-    except KeyError:
-        if sympy_function == sympy.Piecewise:
-            return functions.conditional
-        raise ValueError(f"SymPy function {sympy_function} is not recognized.")
-
-
-def _python_callable_to_sympy_function(python_callable: Callable) -> sympy.Basic:
-    try:
-        return python_callable_sympy_function_dict[python_callable]
-    except KeyError:
-        raise ValueError(f"Python callable {python_callable} is not recognized.")
-
-
-def _is_sympy_value(sympy_object: sympy.Basic) -> bool:
-    return isinstance(sympy_object, sympy.Number) or isinstance(
-        sympy_object, sympy.logic.boolalg.BooleanAtom
-    )
-
-
-def _is_sympy_sum(sympy_object: sympy.Basic) -> bool:
-    return isinstance(sympy_object, sympy.Sum)
-
-
-def _is_sympy_integral(sympy_object: sympy.Basic) -> bool:
-    return isinstance(sympy_object, sympy.Integral)
-
-
-def _build_type_dict(
-    sympy_arguments: SymPyExpression, type_dict: Dict[sympy.Basic, ExpressionType]
-) -> None:
-    update_consistent_dict(type_dict, sympy_arguments.type_dict)
 
 
 def _build_type_dict_from_sympy_arguments(
@@ -182,6 +61,12 @@ def _build_type_dict_from_sympy_arguments(
     return result
 
 
+def _build_type_dict(
+    sympy_arguments: SymPyExpression, type_dict: Dict[sympy.Basic, ExpressionType]
+) -> None:
+    update_consistent_dict(type_dict, sympy_arguments.type_dict)
+
+
 class SymPyExpression(Expression, ABC):
     def __init__(
         self,
@@ -194,6 +79,26 @@ class SymPyExpression(Expression, ABC):
         super().__init__(expression_type)
         self._sympy_object = sympy_object
         self._type_dict = type_dict
+
+    @property
+    def sympy_object(self):
+        return self._sympy_object
+
+    @property
+    def type_dict(self) -> Dict[sympy.Basic, ExpressionType]:
+        return self._type_dict
+
+    def internal_object_eq(self, other) -> bool:
+        match other:
+            case SymPyExpression(
+                sympy_object=other_sympy_object, type_dict=other_type_dict
+            ):
+                return (
+                    self.sympy_object == other_sympy_object
+                    and self.type_dict == other_type_dict
+                )
+            case _:
+                return False
 
     @staticmethod
     def symbolic_sum(
@@ -276,7 +181,7 @@ class SymPyExpression(Expression, ABC):
                 raise FunctionNotTypedError
         else:
             try:
-                sympy_object = _python_callable_to_sympy_function(value)
+                sympy_object = python_callable_to_sympy_function(value)
             except Exception as exc:
                 raise ValueError(
                     f"SymPyConstant does not support {type(value)}: "
@@ -310,7 +215,7 @@ class SymPyExpression(Expression, ABC):
             # if function is not of SymPyConstant but of Constant, then it is assumed to be a python callable
             case Constant(value=python_callable):
                 # during the call, ValueError will be implicitly raised if we cannot convert
-                sympy_function = _python_callable_to_sympy_function(python_callable)
+                sympy_function = python_callable_to_sympy_function(python_callable)
                 return (
                     SymPyFunctionApplication.from_sympy_function_and_general_arguments(
                         sympy_function, arguments
@@ -353,29 +258,9 @@ class SymPyExpression(Expression, ABC):
             return str(value)  # uninterpreted function
         else:
             try:
-                return _sympy_function_to_python_callable(value)
+                return sympy_function_to_python_callable(value)
             except Exception as exc:
                 raise ValueError(f"Cannot pythonize {value}.") from exc
-
-    @property
-    def sympy_object(self):
-        return self._sympy_object
-
-    @property
-    def type_dict(self) -> Dict[sympy.Basic, ExpressionType]:
-        return self._type_dict
-
-    def internal_object_eq(self, other) -> bool:
-        match other:
-            case SymPyExpression(
-                sympy_object=other_sympy_object, type_dict=other_type_dict
-            ):
-                return (
-                    self.sympy_object == other_sympy_object
-                    and self.type_dict == other_type_dict
-                )
-            case _:
-                return False
 
     @staticmethod
     def from_sympy_object(
@@ -384,13 +269,13 @@ class SymPyExpression(Expression, ABC):
         # Here we just try to find a type of expression for sympy object.
         if isinstance(sympy_object, sympy.Symbol):
             return SymPyVariable(sympy_object, type_dict[sympy_object])
-        elif _is_sympy_value(sympy_object):
+        elif is_sympy_value(sympy_object):
             return SymPyConstant(
-                sympy_object, _infer_sympy_object_type(sympy_object, type_dict)
+                sympy_object, infer_sympy_object_type(sympy_object, type_dict)
             )
-        elif _is_sympy_sum(sympy_object):
+        elif is_sympy_sum(sympy_object):
             return SymPySummation(sympy_object, type_dict)
-        elif _is_sympy_integral(sympy_object):
+        elif is_sympy_integral(sympy_object):
             raise NotImplementedError("expect sympy to eliminate integral sign. TODO")
         else:
             return SymPyFunctionApplication(sympy_object, type_dict)
@@ -421,7 +306,7 @@ class SymPyConstant(SymPyExpression, Constant):
         expression_type: Optional[ExpressionType] = None,
     ):
         if expression_type is None:
-            expression_type = _infer_sympy_object_type(sympy_object, {})
+            expression_type = infer_sympy_object_type(sympy_object, {})
         SymPyExpression.__init__(
             self, sympy_object, expression_type, {}
         )  # type_dict only records variables
@@ -433,21 +318,16 @@ class SymPyConstant(SymPyExpression, Constant):
 
 class SymPyFunctionApplicationInterface(SymPyExpression, FunctionApplication, ABC):
     @property
-    @abstractmethod
-    def native_arguments(self) -> Tuple[sympy.Basic, ...]:
-        pass
-
-    @property
-    @abstractmethod
-    def function_type(self) -> ExpressionType:
-        pass
-
-    @property
     def function(self) -> Expression:
         if is_sympy_uninterpreted_function(self._sympy_object.func):
             return SymPyVariable(self._sympy_object.func, self.function_type)
         else:
             return SymPyConstant(self._sympy_object.func, self.function_type)
+
+    @property
+    @abstractmethod
+    def function_type(self) -> ExpressionType:
+        pass
 
     @property
     def arguments(self) -> List[Expression]:
@@ -457,19 +337,16 @@ class SymPyFunctionApplicationInterface(SymPyExpression, FunctionApplication, AB
         ]
 
     @property
+    @abstractmethod
+    def native_arguments(self) -> Tuple[sympy.Basic, ...]:
+        pass
+
+    @property
     def subexpressions(self) -> List[Expression]:
         return [self.function] + self.arguments
 
 
 class SymPyFunctionApplication(SymPyFunctionApplicationInterface):
-    def __new__(
-        cls, sympy_object: sympy.Basic, type_dict: Dict[sympy.Basic, ExpressionType]
-    ):
-        if sympy_object.func == sympy.Piecewise:
-            return SymPyConditionalFunctionApplication(sympy_object, type_dict)
-        else:
-            return super().__new__(cls)
-
     def __init__(
         self, sympy_object: sympy.Basic, type_dict: Dict[sympy.Basic, ExpressionType]
     ):
@@ -486,7 +363,7 @@ class SymPyFunctionApplication(SymPyFunctionApplicationInterface):
             # this happens iff sympy_object is an uninterpreted function, whose type cannot be inferred
             self._function_type = type_dict[sympy_object.func]
         else:
-            self._function_type = _infer_sympy_function_type(sympy_object, type_dict)
+            self._function_type = infer_sympy_function_type(sympy_object, type_dict)
         return_type = return_type_after_application(
             self._function_type, len(sympy_object.args)
         )
@@ -548,35 +425,13 @@ class SymPyFunctionApplication(SymPyFunctionApplicationInterface):
         else:
             return SymPyFunctionApplication(sympy_object, type_dict)
 
-
-def sympy_piecewise_from_if_then_else(
-    if_: sympy.Basic, then_: sympy.Basic, else_: sympy.Basic
-) -> sympy.Piecewise:
-    """In Piecewise, conditional comes after clause."""
-    return sympy.Piecewise((then_, if_), (else_, True))
-
-
-def sympy_piecewise_to_if_then_else(
-    piecewise: sympy.Piecewise,
-) -> Tuple[sympy.Basic, sympy.Basic, sympy.Basic]:
-    return piecewise.args[0][1], piecewise.args[0][0], piecewise.args[1][0]
-
-
-def fold_sympy_piecewise(
-    piecewise_args: List[Tuple[sympy.Basic, sympy.Basic]]
-) -> sympy.Piecewise:
-    """`fold` any sympy piecewise to 2-entry piecewise. E.g.,
-    Piecewise((s0, c0), (s1, c1), (s2, True)) will be folded into
-    Piecewise((s0, c0), (Piecewise((s1, c1), (s2, True)), True)
-    """
-    if len(piecewise_args) < 2:
-        raise TypeError("piecewise is expected to have at least two entries")
-    elif len(piecewise_args) == 2:
-        return sympy.Piecewise(*piecewise_args)
-    else:
-        return sympy.Piecewise(
-            piecewise_args[0], (fold_sympy_piecewise(piecewise_args[1:]), True)
-        )
+    def __new__(
+        cls, sympy_object: sympy.Basic, type_dict: Dict[sympy.Basic, ExpressionType]
+    ):
+        if sympy_object.func == sympy.Piecewise:
+            return SymPyConditionalFunctionApplication(sympy_object, type_dict)
+        else:
+            return super().__new__(cls)
 
 
 class SymPyConditionalFunctionApplication(SymPyFunctionApplicationInterface):
@@ -592,16 +447,16 @@ class SymPyConditionalFunctionApplication(SymPyFunctionApplicationInterface):
         ]:  # the clause condition must be True otherwise it's not an if-then-else
             raise TypeError("Missing else clause.")
         sympy_object = fold_sympy_piecewise(sympy_object.args)
-        self._then_type = _infer_sympy_object_type(sympy_object.args[0][0], type_dict)
+        self._then_type = infer_sympy_object_type(sympy_object.args[0][0], type_dict)
         SymPyExpression.__init__(self, sympy_object, self._then_type, type_dict)
-
-    @property
-    def function_type(self) -> ExpressionType:
-        return Callable[[bool, self._then_type, self._then_type], self._then_type]
 
     @property
     def function(self) -> Expression:
         return SymPyConstant(self._sympy_object.func, self.function_type)
+
+    @property
+    def function_type(self) -> ExpressionType:
+        return Callable[[bool, self._then_type, self._then_type], self._then_type]
 
     @property
     def number_of_arguments(self) -> int:
@@ -610,58 +465,6 @@ class SymPyConditionalFunctionApplication(SymPyFunctionApplicationInterface):
     @property
     def native_arguments(self) -> Tuple[sympy.Basic, ...]:
         return sympy_piecewise_to_if_then_else(self.sympy_object)
-
-
-def _context_to_variable_value_dict_helper(
-    context: FunctionApplication,
-    variable_to_value: Dict[str, Any],
-    unknown: bool = False,
-    unsatisfiable: bool = False,
-) -> Tuple[Dict[str, Any], bool, bool]:
-    """
-    variable_to_value: the mutable argument also serves as a return value.
-    By default, we assume the context's satisfiability can be known and is True.
-    If the context has multiple assignments (e.g., x==3 and x==5), the context is unsatisfiable.
-    If the context is anything other than a conjunction of equalities, the context's satisfiability is unknown.
-    """
-    match context:
-        case FunctionApplication(
-            function=Constant(value=operator.and_), arguments=arguments
-        ):
-            # the conjunctive case
-            for sub_context in arguments:
-                (
-                    variable_to_value,
-                    unknown,
-                    unsatisfiable,
-                ) = _context_to_variable_value_dict_helper(
-                    sub_context, variable_to_value, unknown, unsatisfiable
-                )
-        case FunctionApplication(
-            function=Constant(value=operator.eq),
-            arguments=[Variable(name=variable), Constant(value=value)],
-        ) | FunctionApplication(
-            function=Constant(value=operator.eq),
-            arguments=[Constant(value=value), Variable(name=variable)],
-        ):
-            # the leaf case
-            if variable in variable_to_value and variable_to_value[variable] != value:
-                unsatisfiable = True
-            variable_to_value[variable] = value
-        # all other cases makes the satisfiability unknown
-        case _:
-            unknown = True
-    return variable_to_value, unknown, unsatisfiable
-
-
-def _context_to_variable_value_dict(
-    context: FunctionApplication,
-) -> Tuple[Dict[str, Any], bool, bool]:
-    """
-    Returns a dictionary, and two booleans: first indicating whether its satisfiability is unknown, second indicating
-    whether it is unsatisfiable (if its satisfiability is known)
-    """
-    return _context_to_variable_value_dict_helper(context, {})
 
 
 class SymPyContext(SymPyFunctionApplication, Context):
@@ -678,10 +481,15 @@ class SymPyContext(SymPyFunctionApplication, Context):
             self._dict,
             self._unknown,
             self._unsatisfiable,
-        ) = _context_to_variable_value_dict(self)
+        ) = self._context_to_variable_value_dict(self)
         if self._unsatisfiable:
             # if unsat, looking up dict for value does not make sense, as any value can be detailed
             self._dict = {}
+
+    @property
+    def dict(self) -> Dict[str, Any]:
+        """User should not write to the return value."""
+        return self._dict
 
     @property
     def unsatisfiable(self) -> bool:
@@ -691,26 +499,72 @@ class SymPyContext(SymPyFunctionApplication, Context):
             return self._unsatisfiable
 
     @property
-    def dict(self) -> Dict[str, Any]:
-        """User should not write to the return value."""
-        return self._dict
-
-    @property
     def satisfiability_is_known(self) -> bool:
         return not self._unknown
+
+    def _context_to_variable_value_dict(
+        self,
+        context: FunctionApplication,
+    ) -> Tuple[Dict[str, Any], bool, bool]:
+        """
+        Returns a dictionary, and two booleans: first indicating whether its satisfiability is unknown, second indicating
+        whether it is unsatisfiable (if its satisfiability is known)
+        """
+        return self._context_to_variable_value_dict_helper(context, {})
+
+    def _context_to_variable_value_dict_helper(
+        self,
+        context: FunctionApplication,
+        variable_to_value: Dict[str, Any],
+        unknown: bool = False,
+        unsatisfiable: bool = False,
+    ) -> Tuple[Dict[str, Any], bool, bool]:
+        """
+        variable_to_value: the mutable argument also serves as a return value.
+        By default, we assume the context's satisfiability can be known and is True.
+        If the context has multiple assignments (e.g., x==3 and x==5), the context is unsatisfiable.
+        If the context is anything other than a conjunction of equalities, the context's satisfiability is unknown.
+        """
+        match context:
+            case FunctionApplication(
+                function=Constant(value=operator.and_), arguments=arguments
+            ):
+                # the conjunctive case
+                for sub_context in arguments:
+                    (
+                        variable_to_value,
+                        unknown,
+                        unsatisfiable,
+                    ) = self._context_to_variable_value_dict_helper(
+                        sub_context, variable_to_value, unknown, unsatisfiable
+                    )
+            case FunctionApplication(
+                function=Constant(value=operator.eq),
+                arguments=[Variable(name=variable), Constant(value=value)],
+            ) | FunctionApplication(
+                function=Constant(value=operator.eq),
+                arguments=[Constant(value=value), Variable(name=variable)],
+            ):
+                # the leaf case
+                if (
+                    variable in variable_to_value
+                    and variable_to_value[variable] != value
+                ):
+                    unsatisfiable = True
+                variable_to_value[variable] = value
+            # all other cases makes the satisfiability unknown
+            case _:
+                unknown = True
+        return variable_to_value, unknown, unsatisfiable
 
 
 class SymPySummation(SymPyExpression, QuantifierExpression):
     def __init__(
         self, sympy_object: sympy.Basic, type_dict: Dict[sympy.Basic, ExpressionType]
     ):
-        expression_type = _infer_sympy_object_type(sympy_object.args[0], type_dict)
+        expression_type = infer_sympy_object_type(sympy_object.args[0], type_dict)
         # sympy.Sum(body, (index, lower, upper))
         super().__init__(sympy_object, expression_type, type_dict)
-
-    @property
-    def is_integral(self) -> bool:
-        return False
 
     @property
     def operation(self) -> AbelianOperation:
@@ -720,16 +574,6 @@ class SymPySummation(SymPyExpression, QuantifierExpression):
     def index(self) -> SymPyVariable:
         variable = self.sympy_object.args[1][0]
         return SymPyVariable(variable, self.type_dict[variable])
-
-    @property
-    def lower_bound(self) -> SymPyExpression:
-        lower = self.sympy_object.args[1][1]
-        return SymPyExpression.from_sympy_object(lower, self.type_dict)
-
-    @property
-    def upper_bound(self) -> SymPyExpression:
-        upper = self.sympy_object.args[1][2]
-        return SymPyExpression.from_sympy_object(upper, self.type_dict)
 
     @property
     def constraint(self) -> Context:
@@ -747,3 +591,17 @@ class SymPySummation(SymPyExpression, QuantifierExpression):
         return SymPyExpression.from_sympy_object(
             self.sympy_object.args[0], self.type_dict
         )
+
+    @property
+    def lower_bound(self) -> SymPyExpression:
+        lower = self.sympy_object.args[1][1]
+        return SymPyExpression.from_sympy_object(lower, self.type_dict)
+
+    @property
+    def upper_bound(self) -> SymPyExpression:
+        upper = self.sympy_object.args[1][2]
+        return SymPyExpression.from_sympy_object(upper, self.type_dict)
+
+    @property
+    def is_integral(self) -> bool:
+        return False
