+++ neuralpp/symbolic/sympy_interpreter.py
@@ -2,11 +2,13 @@ import operator
 import sympy
 
 from neuralpp.symbolic.expression import Expression, FunctionApplication, Constant, Variable, Context
-from neuralpp.symbolic.error import ConversionError
+from neuralpp.util.symbolic_error_util import ConversionError
 from neuralpp.symbolic.basic_expression import TrueContext
 from neuralpp.symbolic.simplifier import Simplifier
 from neuralpp.symbolic.interpreter import Interpreter
-from neuralpp.symbolic.sympy_expression import SymPyExpression, _is_sympy_value, _infer_sympy_object_type
+from neuralpp.symbolic.sympy_expression import SymPyExpression
+from neuralpp.util.callable_util import infer_sympy_object_type
+from neuralpp.util.sympy_util import is_sympy_value
 from neuralpp.symbolic.expression import ExpressionType
 from neuralpp.symbolic.parameters import sympy_evaluate
 from typing import Dict, Any
@@ -27,7 +29,7 @@ class SymPyInterpreter(Interpreter, Simplifier):
 
     def eval(self, expression: SymPyExpression, context: Context = TrueContext()):
         result = SymPyInterpreter._simplify_expression(expression.sympy_object, context)
-        if _is_sympy_value(result):
+        if is_sympy_value(result):
             return result
         else:
             raise RuntimeError(f"cannot evaluate to a value. The best effort result is {result}.")
